tests/requirements/requirements-memcached-1.51.txt
python-memcached==1.51
-r requirements-base.txt
